+++ src/components/card/CardAddForm.js
@@ -1,21 +1,21 @@
 import * as Yup from 'yup';
 
-import { BorderRadius, Colors, FontSize, Media, Shadows } from '../styles';
+import { BorderRadius, Colors, FontSize, Media, Shadows } from '../../styles';
 import { Columns, Section } from 'react-bulma-components';
 import { useEffect, useState } from 'react';
 
-import Buttons from './Buttons';
-import IconInput from './IconInput';
-import IconTag from './IconTag';
-import ModalFooter from './modals/ModalFooter';
-import { REACT_QUERY_KEY } from '../constants/query';
-import { cardChangeState } from '../state/cardState';
-import { makeCardFromRequest } from '../utils/cardUtils';
-import { onAddCard } from '../api/cardApi';
+import Buttons from '../common/Buttons';
+import IconInput from '../common/IconInput';
+import IconTag from '../tag/IconTag';
+import ModalFooter from '../modals/ModalFooter';
+import { REACT_QUERY_KEY } from '../../constants/query';
+import { cardChangeState } from '../../state/cardState';
+import { makeCardFromRequest } from '../../utils/cardUtils';
+import { onAddCard } from '../../api/cardApi';
 import styled from '@emotion/styled';
-import useCardChangeWithFolder from '../hooks/useCardChangeWithFolder';
-import useCardChangeWithTag from '../hooks/useCardChangeWithTag';
-import useDebounce from '../hooks/useDebounce';
+import useCardChangeWithFolder from '../../hooks/useCardChangeWithFolder';
+import useCardChangeWithTag from '../../hooks/useCardChangeWithTag';
+import useDebounce from '../../hooks/useDebounce';
 import { useFormik } from 'formik';
 import { useQueryClient } from 'react-query';
 import { useSetRecoilState } from 'recoil';
